+++ relay/internal/testkit/app.go
@@ -7,8 +7,8 @@ import (
 	"net"
 	"time"
 
-	"github.com/dsh-links/dsh-links-relay/internal/cryptoutil"
-	"github.com/dsh-links/dsh-links-relay/internal/protocol"
+	"github.com/lunaship/dsh-links/relay/internal/cryptoutil"
+	"github.com/lunaship/dsh-links/relay/internal/protocol"
 )
 
 // SimApp simulates mobile app